+++ espflash/src/cli/mod.rs
@@ -30,9 +30,10 @@ use crate::{
     connection::reset::{ResetAfterOperation, ResetBeforeOperation},
     elf::ElfFirmwareImage,
     error::{Error, MissingPartition, MissingPartitionTable},
-    flash_data::{parse_partition_table, FlashData, FlashFrequency, FlashMode, FlashSize},
-    flasher::Flasher,
-    progress::ProgressCallbacks,
+    flasher::{
+        parse_partition_table, FlashData, FlashFrequency, FlashMode, FlashSize, Flasher,
+        ProgressCallbacks,
+    },
     targets::{Chip, XtalFrequency},
 };